Relocate `File` and `HttpClient` to their respective modules
